refactor : #390 <Auth> 패키지 구조를 생성
- v1, v2 패키지 구조를 생성하면서 테스트 코드에도 패키지를 생성했습니다.
